requirements.txt
flask
discord.py
line-bot-sdk
ImageHash
pya3rt
pillow
arrow
